chore: move placements to util folder
