+++ components/dropdown/dropdown.tsx
@@ -7,7 +7,7 @@ import { ConfigContext } from '../config-provider';
 import devWarning from '../_util/devWarning';
 import { tuple } from '../_util/type';
 import { cloneElement } from '../_util/reactNode';
-import getPlacements from '../tooltip/placements';
+import getPlacements from '../_util/placements';
 
 const Placements = tuple(
   'topLeft',
